# proto/openapi/seasonal-challenges-service/main.yaml
...
openapi: 3.0.3
info:
  title: Seasonal Challenges Service API
  description: |
    **Enterprise-grade API for Seasonal Challenges System**

    ## Domain Purpose
    Manages seasonal challenges, trials, and competitive events with real-time progression tracking, rewards distribution, and leaderboards for MMOFPS RPG gameplay.

    ## Performance Targets
    - P99 Latency: <50ms for all endpoints
    - Memory per Instance: <50KB baseline
    - Concurrent Users: 10,000+ supported
    - Uptime: 99.9% SLA

    ## Domain Inheritance
    This service inherits from seasonal-challenges-entities.yaml providing:
    - Consistent base entity structure (id, timestamps, version)
    - Domain-specific business entities
    - Optimistic locking for concurrent operations
    - Strict typing with validation rules

  version: "1.0.0"
  contact:
    name: NECPGAME API Team
    email: api@necpgame.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.necpgame.com/v1/seasonal-challenges
    description: Production environment
  - url: https://staging-api.necpgame.com/v1/seasonal-challenges
    description: Staging environment
  - url: http://localhost:8080/api/v1/seasonal-challenges
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /seasons:
    get:
      operationId: listSeasons
      summary: List seasons with filtering and pagination
      description: |
        Retrieve a paginated list of seasons with advanced filtering options.

        **Performance Notes:**
        - Uses database indexes for efficient filtering
        - Implements cursor-based pagination for large datasets
        - Response cached for 30 seconds
      tags: [Seasonal Challenges]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items to return
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
            default: active
          description: Filter by season status
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [start_date, end_date, created_at]
            default: start_date
          description: Sort field
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SeasonSummary'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 150
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    post:
      operationId: createSeason
      summary: Create a new season
      description: |
        Create a new season with validation and business rules.

        **Business Rules:**
        - Validates season dates don't overlap
        - Applies seasonal currency limits
        - Generates audit trail
      tags: [Administration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
      responses:
        '201':
          description: Season created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '../common/responses/success.yaml#/components/responses/Created'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Season'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /ws/events:
    get:
      operationId: connectSeasonalEvents
      summary: WebSocket connection for seasonal events
      description: Establish WebSocket connection for real-time seasonal event notifications
      tags: [Events]
      responses:
        '101':
          description: WebSocket connection established
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  responses:
...
    Created:
      $ref: '../common/responses/success.yaml#/components/responses/Created'
...
    BadRequest:
      $ref: '../common/responses/error.yaml#/components/responses/BadRequest'
    Unauthorized:
      $ref: '../common/responses/error.yaml#/components/responses/Unauthorized'
...
    Conflict:
      $ref: '../common/responses/error.yaml#/components/responses/Conflict'
    TooManyRequests:
      $ref: '../common/responses/error.yaml#/components/responses/TooManyRequests'
...
  schemas:
...
    Season:
      description: Complete season entity
      type: object
      required:
        - id
        - name
        - start_date
        - end_date
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Season ID
        name:
          type: string
          description: Season name
        start_date:
          type: string
          format: date-time
          description: Season start date
        end_date:
          type: string
          format: date-time
          description: Season end date
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        version:
          type: integer
          description: Optimistic locking version
        currency_limit:
          type: integer
          minimum: 0
          maximum: 1000000
          description: Maximum seasonal currency that can be earned
          example: 50000
        rewards_pool:
          type: array
          items:
            $ref: '#/components/schemas/SeasonRewardPool'
          description: Available reward pools for the season
    SeasonSummary:
      description: Summary season for list operations
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Season ID
        name:
          type: string
          description: Season name
        status:
          type: string
          enum: [upcoming, active, completed]
          description: Season status
        start_date:
          type: string
          format: date-time
          description: Season start date
        end_date:
          type: string
          format: date-time
          description: Season end date
        total_challenges:
          type: integer
          description: Total number of challenges
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
...
    Reward:
      description: Seasonal reward entity
      type: object
      properties:
        type:
          type: string
          enum: [currency, item, cosmetic, title, achievement, experience_points, experience_multiplier]
          description: Type of reward
        amount:
          type: integer
          minimum: 1
          description: Reward amount/quantity
        item_id:
          type: string
          format: uuid
          description: Item ID (for item rewards)
        metadata:
          type: object
          description: Additional reward data
    SeasonRewardPool:
      description: Reward pool configuration for season
      type: object
      properties:
        tier:
          type: string
          enum: [bronze, silver, gold, platinum, diamond]
          description: Reward tier
        min_score:
          type: integer
          minimum: 0
          description: Minimum score required for this tier
        rewards:
          type: array
          items:
            $ref: '#/components/schemas/Reward'
          description: Available rewards in this pool
...
    CreateSeasonRequest:
      description: Request payload for creating a season
      type: object
      required:
        - name
        - start_date
        - end_date
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Season name
          example: "Summer Championship 2026"
        description:
          type: string
          maxLength: 500
          description: Season description
        start_date:
          type: string
          format: date-time
          description: Season start date
        end_date:
          type: string
          format: date-time
          description: Season end date
        currency_limit:
          type: integer
          minimum: 0
          maximum: 1000000
          default: 10000
          description: Maximum seasonal currency per player
        reward_pools:
          type: array
          items:
            $ref: '#/components/schemas/SeasonRewardPool'
          description: Reward pools configuration
        metadata:
          type: object
          description: Additional season metadata
...
  securitySchemes:
    BearerAuth:
      $ref: '../common/security/security.yaml#/components/securitySchemes/BearerAuth'
    ApiKeyAuth:
      $ref: '../common/security/security.yaml#/components/securitySchemes/ApiKeyAuth'
    ServiceAuth:
      type: http
      scheme: bearer
      description: Service-to-service authentication

# Performance and optimization notes
...