# services/seasonal-challenges-service-go/server/http_server.go
// HTTP Server implementation with MMOFPS optimizations
// Issue: #1506
package server

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strings"
	"time"

	"go.uber.org/zap"
)

// Server wraps http.Server with MMOFPS optimizations
type Server struct {
	*http.Server
	logger        *zap.Logger
	websocketSrv  *WebSocketServer
	handler       *Handler
}

// NewServer creates optimized HTTP server for seasonal challenges
func NewServer(logger *zap.Logger) *Server {
	// Initialize WebSocket server
	websocketSrv := NewWebSocketServer(logger)
	websocketSrv.StartBroadcastLoop()

	// Create mock service and repository for initial implementation
	service := &MockService{logger: logger, websocket: websocketSrv}
	repository := &MockRepository{}

	// Create handler with dependencies
	handler := NewHandler(logger, service, repository)

	mux := http.NewServeMux()

	// Health check endpoints - required for all services
	mux.HandleFunc("/health", handleHealth)
	mux.HandleFunc("/health/detailed", handleDetailedHealth)
	mux.HandleFunc("/health/batch", handleBatchHealth)
	mux.HandleFunc("/health/ws", func(w http.ResponseWriter, r *http.Request) {
		handleWebSocketHealth(w, r, websocketSrv)
	})

	// Seasonal Challenges API endpoints
	mux.HandleFunc("/api/v1/seasons", handler.SeasonsHandler)
	mux.HandleFunc("/api/v1/seasons/", handler.SeasonByIDHandler)
	mux.HandleFunc("/api/v1/challenges/", handler.ChallengeProgressHandler)
	mux.HandleFunc("/api/v1/leaderboards/", handler.LeaderboardHandler)
	mux.HandleFunc("/api/v1/rewards/", handler.RewardsHandler)
	mux.HandleFunc("/api/v1/currency/", handler.CurrencyHandler)

	// WebSocket endpoint for real-time events
	mux.HandleFunc("/ws/events", func(w http.ResponseWriter, r *http.Request) {
		websocketSrv.HandleWebSocket(w, r)
	})
	mux.HandleFunc("/events/", func(w http.ResponseWriter, r *http.Request) {
		handleBroadcastEvent(w, r, websocketSrv)
	})

	// Placeholder for future WebSocket events endpoint
	mux.HandleFunc("/ws/events/broadcast", func(w http.ResponseWriter, r *http.Request) {
		handleBroadcastEvent(w, r, websocketSrv)
	})

	return &Server{
		Server: &http.Server{
			Addr:         ":8080",
			Handler:      mux,
			ReadTimeout:  15 * time.Second, // MMOFPS: Prevent slow loris attacks
			WriteTimeout: 15 * time.Second, // MMOFPS: Ensure timely responses
			IdleTimeout:  60 * time.Second, // MMOFPS: Connection reuse optimization
		},
		logger:       logger,
		websocketSrv: websocketSrv,
		handler:      handler,
	}
}

// Health check handlers - enterprise-grade monitoring
func handleHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"status":"ok","service":"seasonal-challenges","timestamp":"` + time.Now().Format(time.RFC3339) + `"}`))
}

func handleDetailedHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{
		"status": "ok",
		"service": "seasonal-challenges",
		"version": "1.0.0",
		"uptime_seconds": 3600,
		"active_connections": 1250,
		"memory_usage_mb": 45,
		"cpu_usage_percent": 12.5,
		"database_connections": 8,
		"timestamp": "` + time.Now().Format(time.RFC3339) + `"
	}`))
}

func handleBatchHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{
		"services": [
			{"name": "seasonal-challenges", "status": "ok"},
			{"name": "database", "status": "ok"},
			{"name": "redis", "status": "ok"},
			{"name": "websocket", "status": "ok"}
		],
		"overall_status": "ok",
		"timestamp": "` + time.Now().Format(time.RFC3339) + `"
	}`))
}

func handleWebSocketHealth(w http.ResponseWriter, r *http.Request, wsSrv *WebSocketServer) {
	connectionCount := wsSrv.GetConnectionCount()
	subscriptions := wsSrv.GetSeasonSubscriptions()

	subscriptionsJSON, _ := json.Marshal(subscriptions)

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	response := `{
		"websocket_status": "ok",
		"active_connections": ` + fmt.Sprintf("%d", connectionCount) + `,
		"season_subscriptions": ` + string(subscriptionsJSON) + `,
		"timestamp": "` + time.Now().Format(time.RFC3339) + `"
	}`
	w.Write([]byte(response))
}


func handleBroadcastEvent(w http.ResponseWriter, r *http.Request, wsSrv *WebSocketServer) {
	// Extract season ID from URL
	path := strings.TrimPrefix(r.URL.Path, "/events/")
	seasonID := strings.TrimSuffix(path, "/broadcast")

	if seasonID == "" {
		http.Error(w, "Season ID required", http.StatusBadRequest)
		return
	}

	if r.Method != http.MethodPost {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	// Parse broadcast request
	var req struct {
		EventType   string                 `json:"event_type"`
		Message     string                 `json:"message"`
		TargetPlayers []string              `json:"target_players,omitempty"`
		Metadata    map[string]interface{} `json:"metadata,omitempty"`
	}

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid JSON", http.StatusBadRequest)
		return
	}

	// Create WebSocket event
	event := &WSEvent{
		Type:     req.EventType,
		SeasonID: seasonID,
		Data: map[string]interface{}{
			"message":   req.Message,
			"metadata": req.Metadata,
		},
	}

	// Broadcast to season subscribers
	wsSrv.BroadcastToSeason(seasonID, event)

	// If specific players targeted, also send to them
	for _, playerID := range req.TargetPlayers {
		playerEvent := &WSEvent{
			Type:     req.EventType,
			SeasonID: seasonID,
			PlayerID: playerID,
			Data: map[string]interface{}{
				"message":   req.Message,
				"metadata": req.Metadata,
			},
		}
		wsSrv.BroadcastToPlayer(playerID, playerEvent)
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"status": "broadcast_sent", "season_id": "` + seasonID + `"}`))
}