# services/time-trials-service-go/internal/models/types.go
// Code generated by NECPGAME backend agent. Common models for time trials service.
// Issue: #1508
// Fixes circular import by moving shared types to separate package

package models

import (
	"time"

	"github.com/google/uuid"
)

// TrialConfig represents trial configuration data
type TrialConfig struct {
	ID          uuid.UUID `json:"id" db:"id"`
	TrialID     string    `json:"trial_id" db:"trial_id"`
	Name        string    `json:"name" db:"name"`
	Description string    `json:"description" db:"description"`
	ConfigData  string    `json:"config_data" db:"config_data"` // JSON blob
	IsActive    bool      `json:"is_active" db:"is_active"`
	CreatedAt   time.Time `json:"created_at" db:"created_at"`
	UpdatedAt   time.Time `json:"updated_at" db:"updated_at"`
	CreatedBy   string    `json:"created_by" db:"created_by"`
}

// CompletionResult represents the result of completing a time trial
type CompletionResult struct {
	SessionID          uuid.UUID `json:"session_id"`
	TrialID            uuid.UUID `json:"trial_id"`
	PlayerID           uuid.UUID `json:"player_id"`
	FinalTime          int       `json:"final_time"`          // milliseconds
	ProgressPercentage int       `json:"progress_percentage"` // 0-100
	Status             string    `json:"status"`
	Rank               int       `json:"rank,omitempty"`
	Rewards            []Reward  `json:"rewards,omitempty"`
}

// Reward represents a reward for trial completion
type Reward struct {
	Type     string `json:"type"`     // "experience", "currency", "item"
	Value    int    `json:"value,omitempty"`
	ItemID   string `json:"item_id,omitempty"`
	ItemName string `json:"item_name,omitempty"`
}

// LeaderboardEntry represents a leaderboard entry
type LeaderboardEntry struct {
	PlayerID    uuid.UUID `json:"player_id"`
	PlayerName  string    `json:"player_name"`
	TrialID     uuid.UUID `json:"trial_id"`
	BestTime    int       `json:"best_time"`
	Rank        int       `json:"rank"`
	Achievement string    `json:"achievement,omitempty"`
}

// RewardTier represents reward tier levels
type RewardTier string

const (
	TierBronze   RewardTier = "bronze"
	TierSilver   RewardTier = "silver"
	TierGold     RewardTier = "gold"
	TierPlatinum RewardTier = "platinum"
)

// CurrencyReward represents currency rewards
type CurrencyReward struct {
	Gold           int `json:"gold"`
	SeasonalTokens int `json:"seasonal_tokens"`
}

// ItemReward represents item rewards
type ItemReward struct {
	ItemID   string `json:"item_id"`
	Quantity int    `json:"quantity"`
}

// RewardTiers represents rewards for different tiers
type RewardTiers struct {
	Experience int            `json:"experience"`
	Currency   CurrencyReward `json:"currency"`
	Items      []ItemReward   `json:"items,omitempty"`
	Achievements []string     `json:"achievements,omitempty"`
}

// TrialRewardConfig represents trial reward configuration
type TrialRewardConfig struct {
	BronzeTime      int         `json:"bronze_time"`
	SilverTime      int         `json:"silver_time"`
	GoldTime        int         `json:"gold_time"`
	PlatinumTime    int         `json:"platinum_time"`
	BronzeRewards   RewardTiers `json:"bronze_rewards"`
	SilverRewards   RewardTiers `json:"silver_rewards"`
	GoldRewards     RewardTiers `json:"gold_rewards"`
	PlatinumRewards RewardTiers `json:"platinum_rewards"`
}

// RewardItem represents an item reward
type RewardItem struct {
	ID   string `json:"id"`
	Name string `json:"name"`
	Rarity string `json:"rarity,omitempty"`
}